+++ pevLib/logger.cpp
@@ -21,35 +21,35 @@ logger_class& logger_class::operator<<(const std::wstring& rhs)
 	char *toWrite;
 	if (useWriteConsole)
 	{
-		WriteConsole(stdOut,rhs.c_str(),rhs.length(),&squat,NULL);
+		WriteConsole(stdOut,rhs.c_str(),static_cast<DWORD>(rhs.length()),&squat,NULL);
 		return *this;
 	}
 	switch(globalOptions::encoding)
 	{
 	case globalOptions::ENCODING_TYPE_ACP:
-		len = WideCharToMultiByte(CP_ACP,WC_NO_BEST_FIT_CHARS,rhs.c_str(),rhs.length(),NULL,NULL,"?",&blank);
+		len = WideCharToMultiByte(CP_ACP,WC_NO_BEST_FIT_CHARS,rhs.c_str(),static_cast<int>(rhs.length()),NULL,NULL,"?",&blank);
 		toWrite = new char[len];
-		WideCharToMultiByte(CP_ACP,WC_NO_BEST_FIT_CHARS,rhs.c_str(),rhs.length(),toWrite,len,"?",&blank);
+		WideCharToMultiByte(CP_ACP,WC_NO_BEST_FIT_CHARS,rhs.c_str(),static_cast<int>(rhs.length()),toWrite,len,"?",&blank);
 		WriteFile(stdOut,toWrite,len,&squat,NULL);
 		delete [] toWrite;
 		break;
 	case globalOptions::ENCODING_TYPE_OEM:
-		len = WideCharToMultiByte(CP_OEMCP,WC_NO_BEST_FIT_CHARS,rhs.c_str(),rhs.length(),NULL,NULL,"?",&blank);
+		len = WideCharToMultiByte(CP_OEMCP,WC_NO_BEST_FIT_CHARS,rhs.c_str(),static_cast<int>(rhs.length()),NULL,NULL,"?",&blank);
 		toWrite = new char[len];
-		WideCharToMultiByte(CP_OEMCP,WC_NO_BEST_FIT_CHARS,rhs.c_str(),rhs.length(),toWrite,len,"?",&blank);
+		WideCharToMultiByte(CP_OEMCP,WC_NO_BEST_FIT_CHARS,rhs.c_str(),static_cast<int>(rhs.length()),toWrite,len,"?",&blank);
 		WriteFile(stdOut,toWrite,len,&squat,NULL);
 		delete [] toWrite;
 		break;
 	case globalOptions::ENCODING_TYPE_UTF8:
-		len = WideCharToMultiByte(CP_UTF8,NULL,rhs.c_str(),rhs.length(),NULL,NULL,NULL,NULL);
+		len = WideCharToMultiByte(CP_UTF8,NULL,rhs.c_str(),static_cast<int>(rhs.length()),NULL,NULL,NULL,NULL);
 		toWrite = new char[len];
-		WideCharToMultiByte(CP_UTF8,NULL,rhs.c_str(),rhs.length(),toWrite,len,NULL,NULL);
+		WideCharToMultiByte(CP_UTF8,NULL,rhs.c_str(),static_cast<int>(rhs.length()),toWrite,len,NULL,NULL);
 		WriteFile(stdOut,toWrite,len,&squat,NULL);
 		delete [] toWrite;
 		break;
 	case globalOptions::ENCODING_TYPE_UTF16:
 	default:
-		WriteFile(stdOut,rhs.c_str(),rhs.length()*sizeof(TCHAR),&squat,NULL);
+		WriteFile(stdOut,rhs.c_str(),static_cast<int>(rhs.length())*sizeof(TCHAR),&squat,NULL);
 		break;
 	}
 	return *this;
